+++ docs/components/plugins/api/APISectionComponents.tsx
@@ -15,6 +15,7 @@ import {
   STYLES_APIBOX,
   getTagNamesList,
   H3Code,
+  getPossibleComponentPropsNames,
 } from '~/components/plugins/api/APISectionUtils';
 import { H2, DEMI, P, CODE, MONOSPACE } from '~/ui/components/Text';
 
@@ -83,7 +84,7 @@ const renderComponent = (
         <APISectionProps
           sdkVersion={sdkVersion}
           data={componentsProps}
-          header={componentsProps.length === 1 ? 'Props' : `${resolvedName}Props`}
+          header={`${resolvedName}Props`}
         />
       ) : null}
     </div>
@@ -99,7 +100,7 @@ const APISectionComponents = ({ data, sdkVersion, componentsProps }: APISectionC
           component,
           sdkVersion,
           componentsProps.filter(cp =>
-            cp.name.includes(getComponentName(component.name, component.children))
+            getPossibleComponentPropsNames(component.name, component.children).includes(cp.name)
           )
         )
       )}